ContributionAverageとHatenaCountAverageのrawデータの型変更

+++ app/domain/qiita/user/contribution/ContributionAverage.scala
@@ -3,4 +3,4 @@ package domain.qiita.user.contribution
 /**
   * 一記事あたりのContribution数
   */
-final case class ContributionAverage(value: Float)
+final case class ContributionAverage(value: Int)
