+++ app/domain/qiita/user/contribution/HatenaCountAverage.scala
@@ -3,4 +3,4 @@ package domain.qiita.user.contribution
 /**
   * 一記事あたりのはてなブックマーク数
   */
-final case class HatenaCountAverage(value: Float)
+final case class HatenaCountAverage(value: Int)
